# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Headless E-commerce API",
    "version": "1.0.0",
    "description": "A comprehensive headless e-commerce backend API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/cart/{token}/promo": {
      "post": {
        "summary": "Apply promo code to cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promo code applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired promo code"
          },
          "404": {
            "description": "Cart not found"
          }
        }
      },
      "delete": {
        "summary": "Remove promo code from cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
...
    "/api/promos": {
      "get": {
        "summary": "Get promo codes with filtering and pagination",
        "tags": [
          "Promos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "percentage",
                "fixed"
              ]
            },
            "description": "Filter by promo type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of promo codes with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "promoCodes": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PromoCode"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new promo code",
        "tags": [
          "Promos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoCodeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Promo code created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PromoCode"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "409": {
            "description": "Promo code already exists"
          }
        }
      }
    },
    "/api/promos/{id}": {
      "get": {
        "summary": "Get promo code by ID",
        "tags": [
          "Promos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promo code ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PromoCode"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promo code not found"
          }
        }
      },
      "put": {
        "summary": "Update promo code",
        "tags": [
          "Promos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promo code ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoCodeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promo code updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PromoCode"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Promo code not found"
          }
        }
      },
      "delete": {
        "summary": "Delete promo code",
        "tags": [
          "Promos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promo code ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promo code not found"
          }
        }
      }
    },
    "/api/promos/validate/{code}": {
      "get": {
        "summary": "Validate promo code",
        "tags": [
          "Promos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promo code to validate"
          },
          {
            "in": "query",
            "name": "subtotal",
            "schema": {
              "type": "number"
            },
            "description": "Order subtotal for validation"
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isValid": {
                          "type": "boolean"
                        },
                        "discount": {
                          "type": "number"
                        },
                        "promoCode": {
                          "$ref": "#/components/schemas/PromoCode"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid promo code"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page"
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "pages": {
            "type": "integer",
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "description": "Whether there is a previous page"
          }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Cart ID"
          },
          "token": {
            "type": "string",
            "description": "Cart token"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "quantity": {
                  "type": "integer"
                },
                "subtotal": {
                  "type": "number"
                }
              }
            }
          },
          "subtotal": {
            "type": "number",
            "description": "Cart subtotal"
          },
          "discount": {
            "type": "number",
            "description": "Applied discount amount"
          },
          "total": {
            "type": "number",
            "description": "Cart total after discount"
          },
          "promoCode": {
            "type": "string",
            "description": "Applied promo code"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Cart expiration date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PromoCode": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Promo code ID"
          },
          "code": {
            "type": "string",
            "description": "Promo code"
          },
          "type": {
            "type": "string",
            "enum": [
              "percentage",
              "fixed"
            ],
            "description": "Discount type"
          },
          "value": {
            "type": "number",
            "description": "Discount value"
          },
          "minOrderAmount": {
            "type": "number",
            "description": "Minimum order amount required"
          },
          "maxUses": {
            "type": "integer",
            "description": "Maximum number of uses"
          },
          "usedCount": {
            "type": "integer",
            "description": "Number of times used"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the promo code is active"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}